P1S13 Wizards: fix objectives not changing correctly
This was some combination of the [show_if]s not working correctly when
the objectives were set for several sides simultaneously and delayed
variable substitution (or not).  Not 100% sure exactly whether there is
a Wesnoth bug at play here or not, but I think what I have no should
work.