P2 BRITTLE_EVENT macro: simplify the side filter
It is easier and makes more sense to use it for the side it affects,
rather than for the side that is Grnk’s enemy.  The previous was done
to make it easier in S14 and S15, which have many undead sides.
However, as it is not used there any more, this is more logical.