P2S12 Alliance: small change to the victory conditions
Now we only need to move a unit next to Karcyn, rather than having to
get a hit in.  This is done mostly so that XP for that last hit don’t
need to be artificially carried over to S12s (they did without problems
to S13 previously already), and because it makes sense from a story
point of view anyway.